Fixing the bug :D

+++ src/main/java/com/zgamelogic/controllers/DataOtterController.java
@@ -52,7 +52,6 @@ public class DataOtterController {
         log.info("Performing initial cleanup");
         cleanup();
         log.info("Cleanup completed");
-        System.out.println(monitorStatusRepository.findTopById_Monitor_Id_MonitorConfigurationIdAndId_Monitor_Id_Application_IdOrderById_DateDesc(1L, 1L));
     }
 
     /**
@@ -103,7 +102,9 @@ public class DataOtterController {
                 if(previousStatus.isStatus() != monitorStatus.isStatus()) {
                     changedMonitors.add(monitorStatus);
                 } else {
-                    monitorStatusRepository.deleteById_Monitor_Id_Application_IdAndId_Monitor_Id_MonitorConfigurationIdAndId_Date(previousStatus.getId().getMonitor().getId().getApplication().getId(), previousStatus.getId().getMonitor().getId().getMonitorConfigurationId(), previousStatus.getId().getDate());
+                    if(monitorStatusRepository.countById_Monitor_Id_MonitorConfigurationIdAndId_Monitor_Id_Application_Id(previousStatus.getId().getMonitor().getId().getApplication().getId(), previousStatus.getId().getMonitor().getId().getMonitorConfigurationId()) > 1) {
+                        monitorStatusRepository.deleteById_Monitor_Id_Application_IdAndId_Monitor_Id_MonitorConfigurationIdAndId_Date(previousStatus.getId().getMonitor().getId().getApplication().getId(), previousStatus.getId().getMonitor().getId().getMonitorConfigurationId(), previousStatus.getId().getDate());
+                    }
                 }
             });
             monitorStatusRepository.save(monitorStatus);
